продолжаю выполнять задачу с profileImage
